Create Target.py
Let's consider a standard darts target that consists of K concentric circles with the corresponding radiuses R1,R2,...,Rk with the common center in the origin (0,0).

If your shot lands inside the smallest circle, you will get K points. Landing between the ith and the (i+1)th circle will give you i points. This means your shot includes the ith circle, but excludes the (i+1)th circle. If the shot lands on the boundary of the circle, it will be considered to have landed inside that circle.

Finally, if you are unable to land inside or on the boundary of the 1st circle, you will get 0 points for that shot.

You are given coordinates xi, yi of N shots. Calculate the final score (the sum of all the points).

Input Format

The first line contains two space-separated integers: K and N.

The second line contains K space-separated integers: R1 , R2 ,...Rk.

The following N lines contain two-space separated integers xi, yi, the coordinates of the ith shot.

Constraints

1 ≤ K ≤ 10^4
1 ≤ N ≤ 5 × 10^5
1 ≤ Rk< R k−1 <...< R1 ≤ 5 × 10^4
|xi| , |yi| ≤ 5 × 10^4